+++ src/apis/notice/schema.ts
@@ -8,7 +8,24 @@ import {
   shopSchema,
 } from "@/apis/schema";
 
-export const noticesResponseSchema = z
+export const noticesPostResponseSchema = z
+  .object({
+    item: noticeSchema.merge(
+      z.object({ shop: z.object({ item: shopSchema }).merge(hrefSchema) }),
+    ),
+  })
+  .merge(linksSchema);
+
+export type NoticesPostResponse = z.infer<typeof noticesPostResponseSchema>;
+
+export type NoticesPostRequestBody = {
+  hourlyPay: number;
+  startsAt: string;
+  workhour: number;
+  description: string;
+};
+
+export const noticesGetResponseSchema = z
   .object({
     items: z.array(
       z
@@ -29,3 +46,4 @@ export const noticesResponseSchema = z
       keyword: z.string().optional(),
     }),
   );
+export type NoticesGetResponse = z.infer<typeof noticesGetResponseSchema>;
